The addToFav() and addToCart() methods were moved to FavAndCart.js file. Then, the ProductDetail component was updated with the add to favorites and add to cart buttons.
