+++ src/App.js
@@ -16,18 +16,39 @@ import Favorites from "./components/Favorites/Favorites";
 import CookieCard from './components/CookieCard';
 import Dashboard from './components/Administrator/admin';
 import AlertMessage from "./components/Alerts/AlertMessage";
-import React from "react";
+import React, {useState} from "react";
 import {useSelector} from "react-redux";
 import StripeComponent from './components/StripeComponent/StripeComponent';
 import ShoopingCart from "./components/ShoppingCart";
 import UserControl from './components/UserControl/UserControl';
 import UserDetail from './components/UserControl/UserDetail';
 import UserEdit from './components/UserControl/UserEdit';
-import UserEditData from './components/UserProfile/UserEditData';
+import {Alert, AlertTitle, Snackbar} from "@mui/material";
 
 function App() {
 
     const alertInfo = useSelector(store => store.alertInfo)
+    //-------------------------------
+    //-------------------------------
+    // alert para los fav y cart
+    const [added, setAdded] = useState(false);
+    const [notAdded, setNotAdded] = useState(false);
+
+    const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+            return;
+        }
+        setAdded(false);
+        setNotAdded(false)
+    };
+    const handleAdded = () => {
+        setAdded(true)
+    }
+    const handleNotAdded = () => {
+        setNotAdded(true)
+    }
+    //-------------------------------
+    //-------------------------------
 
     return (<>
         <Router>
@@ -36,9 +57,9 @@ function App() {
             <AlertMessage {...alertInfo} />
             <Routes>
                 <Route exact path="/" element={<LandingPage/>}/>
-                <Route path="/home" element={<Home/>}/>
+                <Route path="/home" element={<Home handleAdded={handleAdded} handleNotAdded={handleNotAdded}/>}/>
                 <Route exact path='/contact' element={<ContactUs/>}/>
-                <Route path='/detail/:id' element={<ProductDetail/>}/>
+                <Route path='/detail/:id' element={<ProductDetail handleAdded={handleAdded} handleNotAdded={handleNotAdded}/>}/>
                 <Route path='/edit/:id' element={<ProductEdit/>}/>
                 <Route exact path='/create' element={<CreateProduct/>}/>
                 <Route exact path='/profile/*' element={<UserProfile/>}/>
@@ -53,6 +74,18 @@ function App() {
                 <Route path='/stripe' element={<StripeComponent/>}/>
                 <Route path='*' element={<NotFound/>}/>
             </Routes>
+            <Snackbar open={added} autoHideDuration={1000} onClose={handleClose}>
+                <Alert onClose={() => handleClose()} severity="success" sx={{ width: '100%' }}>
+                    <AlertTitle>Success</AlertTitle>
+                    <strong>Added correctly</strong>
+                </Alert>
+            </Snackbar>
+            <Snackbar open={notAdded} autoHideDuration={1000} onClose={handleClose}>
+                <Alert onClose={() => handleClose()} severity="warning" sx={{ width: '100%' }}>
+                    <AlertTitle>Fail</AlertTitle>
+                    <strong>List maximum size exceeded</strong>
+                </Alert>
+            </Snackbar>
             <Footer/>
         </Router>
     </>
