+++ src/views/ProductDetail/ProductDetail.css
@@ -100,6 +100,7 @@
 .share-favorite {
     display: flex;
     justify-content: center;
+    align-items: center;
     margin-bottom: 2.5em;
     font-size: .8em;
     gap: 1em;
